+++ src/translations/translations.fr.js
@@ -414,4 +414,31 @@ export const translations = {
   // cc-zone-input
   'cc-zone-input.error': `Une erreur est survenue pendant le chargement des zones.`,
   'cc-zone-input.private-map-warning': `Les zones privées n'apparaissent pas sur la carte.`,
+  // cc-pricing-table
+  'cc-pricing-table.add-button': 'Ajouter',
+  'cc-pricing-table.plan': `Plan`,
+  'cc-pricing-table.price': ({ price, code }) => formatCurrency(lang, price, { currency: code }),
+  'cc-pricing-table.price-name-daily': `Prix (jour)`,
+  'cc-pricing-table.price-name-monthly': () => sanitize`Prix (30&nbsp;jours)`,
+  // -> Feature name translation
+  'cc-pricing-table.feature.connection-limit': `Limite de connexions`,
+  'cc-pricing-table.feature.cpu': `vCPUs`,
+  'cc-pricing-table.feature.databases': `Bases de données`,
+  'cc-pricing-table.feature.disk-size': `Taille du disque`,
+  'cc-pricing-table.feature.gpu': `GPUs`,
+  'cc-pricing-table.feature.has-logs': `Logs`,
+  'cc-pricing-table.feature.has-metrics': `Métriques`,
+  'cc-pricing-table.feature.max-db-size': `Taille BDD max`,
+  'cc-pricing-table.feature.memory': `RAM`,
+  'cc-pricing-table.feature.version': `Version`,
+  // -> Feature type format
+  'cc-pricing-table.type.boolean': ({ boolean }) => `${boolean ? 'Oui' : 'Non'}`,
+  'cc-pricing-table.type.boolean-shared': ({ shared }) => `${shared ? 'Partagé' : 'Dédié'}`,
+  'cc-pricing-table.type.bytes': ({ bytes }) => formatBytes(bytes, 0, 3),
+  'cc-pricing-table.type.number': ({ number }) => formatNumber(lang, number),
+  'cc-pricing-table.type.number-cpu-runtime': ({ cpu, shared }) => {
+    return shared
+      ? sanitize`<em title="Accès au vCPU moins prioritaire">${formatNumber(lang, cpu)}<code>*</code></em>`
+      : formatNumber(lang, cpu);
+  },
 };